crypto_monitor/requirements.txt
aiohttp==3.9.1
aiohttp-socks==0.8.4
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4
plotly==5.18.0
python-socks==2.4.4
playwright==1.41.0
pytest-asyncio==0.23.3
pytest-playwright==0.4.3
pytest==7.4.4
asynctest==0.13.0
python-binance==1.0.19
cryptography==41.0.7
requests==2.31.0
loguru==0.7.2
